Add production configuration to Amplify backend
This commit adds configuration data for a production environment in 'team-provider-info.json' and updates the 'AmplifyDependentResourcesAttributes' type in 'amplify-dependent-resources-ref.d.ts' in Amplify backend. The production config includes information related to cloudformation, authentication roles, region, deployment bucket and more. These changes are necessary to ensure the application works correctly in a production environment.