Reconfigure Cypress setup and TypeScript
This commit removes a number of Cypress setup files and re-adds some with changes, in order to align with the latest project requirements. Additionally, it modifies 'tsconfig.json' to switch to ES2020 for module and target. This aligns the project with the latest JavaScript standard, improving compatibility and potential performance.